Updating the domain to allow country select

+++ plugins/harassmap/incidents/updates/builder_table_create_harassmap_incidents_city.php
@@ -14,7 +14,7 @@ class BuilderTableCreateHarassmapIncidentsCity extends Migration
             $table->string('name', 255);
             $table->string('lat', 255);
             $table->string('lng', 255);
-            $table->smallInteger('country_id')->unsigned();
+            $table->integer('country_id')->unsigned();
 
             $table->foreign('country_id')->references('id')->on('harassmap_incidents_country');
         });
